Adds more flexibility to the foreman:export task
It allows for specifying a custom location for the upstart configuration
files and also allows for setting the use that will be used to run the
foreman export command on the remote server.

Both options make it possible to create a directory inside /etc/initd for
the application being deployed, owned by the deployment user and thus
removing the need for the user to be a sudoer. By combining this with
the ability to run specific foreman commands as sudo, we won't need at
all to be able to become root for this to work.